requirements.txt
langchain
langchain-core
langchain-groq
python-dotenv
GitPython
streamlit
